+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/visitors/SessionActionVisitor.java
@@ -1,13 +1,13 @@
 package org.batfish.datamodel.visitors;
 
-import org.batfish.datamodel.flow.AcceptVrf;
+import org.batfish.datamodel.flow.Accept;
 import org.batfish.datamodel.flow.FibLookup;
 import org.batfish.datamodel.flow.ForwardOutInterface;
 
 /** A generic {@link org.batfish.datamodel.flow.SessionAction} visitor. */
 public interface SessionActionVisitor<T> {
 
-  T visitAcceptVrf(AcceptVrf acceptVrf);
+  T visitAcceptVrf(Accept acceptVrf);
 
   T visitFibLookup(FibLookup fibLookup);
